Implement network::handle_upload()
This function, upon receiving a package upload request from the client,
checks whether the package firstly even exists in the network and if it
does, does the network have an older version of it. In those cases, the
package is accepted and when the client has finished uploading the package
to IPFS, it sends the IPFS hash to the server which then updates the hashmap
of all available packages.

If the version that is being uploaded matches the one that the network
has or is older, the package is rejected.

I'm currently not too happy with the implementation. Using both hashmaps
and vectors, doing unnecessary(?) memory copies and using one
Request/Response message pair for all messages types is sloppy design.