Implement network::handle_update()
The server listens to incoming Update requests from clients. The request
contains a list of packages the client has and as the server has the most
up to date information on all available packages, it is able to check if
there is a newer version available for a package that the clien has.

The response contains a list of packages which have a newer version
available and the client can then update their packages using this list.