Implement network::update()
This function constructs a list from all packages available on the
computer, sends the list to the server and receives a list of packages
that are newer version of the ones that are on the computer.

Using this list, the client then updates all packages which have
a newer version available.

The code requires refactoring because network::update() calls
network::download() which calls network::query() but this querying
is an unnecessary step because the IPFS hash is known by network::update()