Implement querying functionality for the new architecture
Instead of having one server which handles all query, update, add,
and download requests, use IPFS's Publish/Subscribe to handle all
these requests in a decentralized manner.

There are two pubsub interfaces, one for queries and and for responses.
Anybody can query packages but anyone can also responds to them,
removing the single point of failure that was present in the previous
implementation.

The architecture makes the implementation more robust as there are
multiple serving nodes but also much cleaner as the pubsub abstracts
away a lot of the "unnecessary" socket-level stuff that was used
in the previous implementation.