Implement --update-keyring
This option is used to query possible keyring updates. It can be done,
for example, if the end-user notices that a lot of packages that he
downloads gets rejected due to signature mismatch. This is caused
either because his keyring is not up to date and it rejects packages
from a new maintainer, or there is a malicious package in the network
in which case the package handling works as expected.

All of the signer records contain a signature which is calculated
by the initial node of the system and the end-user can ascertain
that the maintainer is trustworthy by checking the signature which
validates the public key of the maintainer.